refactor: move utils to utils folder
